requirements.txt
crewai
python-dotenv
